refactor(sedona-gdal): remove raster re-export aliases
