+++ c/sedona-gdal/src/raster/rasterize.rs
@@ -26,7 +26,7 @@ use crate::dataset::Dataset;
 use crate::errors::{GdalError, Result};
 use crate::gdal_api::{call_gdal_api, GdalApi};
 use crate::gdal_dyn_bindgen::*;
-use crate::vector::Geometry;
+use crate::vector::geometry::Geometry;
 
 /// Source of burn values.
 #[derive(Copy, Clone, Debug)]
